Emily's first pass
I took a first pass and left a few comments set of by <!--EM-->. One concern is the complexity of the code pattern. I was tracking with it until we added information about Watson Translation and Watson Natural Language Unerstanding. Do you think those technologies are important to discuss in this pattern or is it forcing too much into one pattern? Let me know. If we're going to keep it, I think we will need to rethink the Description section to talk about translation.